+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/HiddenBluetoothGattCallback.java
@@ -1,0 +1,11 @@
+package com.polidea.rxandroidble2;
+
+import android.bluetooth.BluetoothGatt;
+
+/**
+ * An interface that represents methods hidden in {@link android.bluetooth.BluetoothGattCallback}
+ */
+public interface HiddenBluetoothGattCallback {
+
+    void onConnectionUpdated(BluetoothGatt gatt, int interval, int latency, int timeout, int status);
+}