+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/internal/operations/OperationsProviderImpl.java
@@ -3,7 +3,6 @@ package com.polidea.rxandroidble2.internal.operations;
 import android.bluetooth.BluetoothGatt;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattDescriptor;
-import android.os.Build;
 import androidx.annotation.RequiresApi;
 
 import com.polidea.rxandroidble2.ClientComponent;
@@ -69,7 +68,7 @@ public class OperationsProviderImpl implements OperationsProvider {
     }
 
     @Override
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @RequiresApi(21 /* Build.VERSION_CODES.LOLLIPOP */)
     public MtuRequestOperation provideMtuChangeOperation(int requestedMtu) {
         return new MtuRequestOperation(rxBleGattCallback, bluetoothGatt, timeoutConfiguration, requestedMtu);
     }
@@ -107,7 +106,7 @@ public class OperationsProviderImpl implements OperationsProvider {
     }
 
     @Override
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @RequiresApi(21 /* Build.VERSION_CODES.LOLLIPOP */)
     public ConnectionPriorityChangeOperation provideConnectionPriorityChangeOperation(int connectionPriority,
                                                                                       long delay,
                                                                                       TimeUnit timeUnit) {