+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/internal/scan/AndroidScanObjectsConverter.java
@@ -2,7 +2,6 @@ package com.polidea.rxandroidble2.internal.scan;
 
 
 import android.annotation.SuppressLint;
-import android.os.Build;
 import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 import androidx.annotation.RestrictTo;
@@ -26,10 +25,10 @@ public class AndroidScanObjectsConverter {
 
     // TODO [DS 18.05.2017] Consider a different implementation for Marshmallow
     @SuppressLint("NewApi")
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @RequiresApi(21 /* Build.VERSION_CODES.LOLLIPOP */)
     public android.bluetooth.le.ScanSettings toNativeSettings(ScanSettings scanSettings) {
         final android.bluetooth.le.ScanSettings.Builder builder = new android.bluetooth.le.ScanSettings.Builder();
-        if (deviceSdk >= Build.VERSION_CODES.M) {
+        if (deviceSdk >= 23 /* Build.VERSION_CODES.M */) {
             setMarshmallowSettings(scanSettings, builder);
         }
         return builder
@@ -38,7 +37,7 @@ public class AndroidScanObjectsConverter {
                 .build();
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
+    @RequiresApi(23 /* Build.VERSION_CODES.M */)
     private void setMarshmallowSettings(ScanSettings scanSettings, android.bluetooth.le.ScanSettings.Builder builder) {
         builder
                 .setCallbackType(scanSettings.getCallbackType())
@@ -47,7 +46,7 @@ public class AndroidScanObjectsConverter {
     }
 
     @Nullable
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @RequiresApi(21 /* Build.VERSION_CODES.LOLLIPOP */)
     public List<android.bluetooth.le.ScanFilter> toNativeFilters(ScanFilter... scanFilters) {
         final boolean isFilteringDefined = scanFilters != null && scanFilters.length > 0;
         final List<android.bluetooth.le.ScanFilter> returnList;
@@ -62,7 +61,7 @@ public class AndroidScanObjectsConverter {
         return returnList;
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @RequiresApi(21 /* Build.VERSION_CODES.LOLLIPOP */)
     private android.bluetooth.le.ScanFilter toNative(ScanFilter scanFilter) {
         final android.bluetooth.le.ScanFilter.Builder builder = new android.bluetooth.le.ScanFilter.Builder();
         if (scanFilter.getServiceDataUuid() != null) {