+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/internal/scan/ScanSetupBuilderImplApi21.java
@@ -1,7 +1,6 @@
 package com.polidea.rxandroidble2.internal.scan;
 
 
-import android.os.Build;
 import androidx.annotation.RequiresApi;
 import androidx.annotation.RestrictTo;
 
@@ -36,7 +35,7 @@ public class ScanSetupBuilderImplApi21 implements ScanSetupBuilder {
         this.androidScanObjectsConverter = androidScanObjectsConverter;
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @RequiresApi(21 /* Build.VERSION_CODES.LOLLIPOP */)
     @Override
     public ScanSetup build(ScanSettings scanSettings, ScanFilter... scanFilters) {
         /*